requirements.txt
# ─────────────────────────────────────────────
# AI Data Analyst Agent — Dependencies
# ─────────────────────────────────────────────

# Orchestration & LLM
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# Data Stack
pandas>=2.2.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Frontend
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0

# RAG & Document Search
faiss-cpu>=1.8.0
pypdf>=4.2.0
langchain-community>=0.2.0
